backend/pyproject.toml
[tool.poetry]
name = "legalops-backend"
version = "0.1.0"
description = "LegalOps SaaS — backend API"
authors = ["You <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }
asyncpg = ">=0.29.0"
psycopg2-binary = "^2.9.0"
alembic = "^1.13.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
bcrypt = "^4.1.0"
httpx = "^0.27.0"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.131.0"
celery = { extras = ["redis"], version = "^5.4.0" }
redis = "^5.0.0"
pypaystack2 = "^3.0.0"
python-multipart = "^0.0.9"
cryptography = "^42.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
anyio = "^4.4.0"
ruff = "^0.4.0"
mypy = "^1.10.0"
pytest-html = "^4.1.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--html=report.html --self-contained-html"

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"